% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/allStations.R
\name{allStations}
\alias{allStations}
\title{Get full list of NOAA ISD stations}
\usage{
allStations()
}
\value{
This function will download the full list of NOAA ISD stations to a variable.
}
\description{
This function is included in case the list of stations is updated.  The full list of stations
is included in the package under the name 'station.list'.
}
\examples{
\dontrun{

station.list <- allStations()
}
}

