% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{api_icons}
\alias{api_icons}
\title{Weather Icon API matches}
\format{A \code{data.frame} with 343 rows and 2 variables:
\describe{
  \item{api}{API icons names}
  \item{icon}{Name of the matching icon}
}}
\description{
Weather Icon API matches
}
\keyword{data}
