% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_weather_stations.R
\name{available_weather_stations}
\alias{available_weather_stations}
\title{Download available weather stations}
\usage{
available_weather_stations(file_dir = "", save_file = TRUE)
}
\arguments{
\item{file_dir}{Directory to save the CSV file. If empty, a temporary directory is created.}

\item{save_file}{Logical indicating whether to save the CSV file. If `FALSE``, the CSV file is removed after reading. Default is `TRUE``.}
}
\value{
data frame of available weather stations.
}
\description{
Downloads all available weather stations from NOAA (around 29,000).

## Columns

- `station_id`: NOAA's two digit country code
- `name`: country name (English)
- `country`: FIPS 10-4 country codes
- `state`: US state (two letter abbreviation) if applicable
- `icao`: International Civil Aviation Organization (ICAO) identifier
- `lat`: latitude
- `lon`: longitude
- `elev`: elevation (meters)
- `begin_date`: first date of operation
- `end_date`: last date of operation
}
\examples{
# Download all available weather stations from NOAA.
stations <- available_weather_stations()

}
