% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_ombrothermic_diagram.R
\name{c_ombrothermic_diagram}
\alias{c_ombrothermic_diagram}
\title{c_ombrothermic_diagram}
\usage{
c_ombrothermic_diagram(data, year_start, year_end, source)
}
\arguments{
\item{data}{Dataframe with the columns \code{date}, \code{rain}, \code{T_air_max} and
\code{T_air_min} respectively the date of the day, the rain (mm) and the maximum
and minimum temperature (°C).}

\item{year_start}{Start year to begin the statical analysis.}

\item{year_end}{End year to finish the statical analysis.}

\item{source}{Source of the data.}
}
\value{
List with \code{table} and \code{diagram}. The table is used to build the
ombrothermic plot.
}
\description{
Plot a ombrothermic diagram for a set of weather data.

On the y scale, one temperature unit equals two rain unit. To
be correct, 30 years must be selected for the statistical analysis.
}
\examples{
output <- c_ombrothermic_diagram(data = weather_blue_grass_airport,
                                 year_start = 1990,
                                 year_end = 2019,
                                 source = "Blue Grass Airport (Lexington)")
}
