% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudmap.R
\name{cloudmap}
\alias{cloudmap}
\title{distmap(api, city1, city2)}
\usage{
cloudmap(api, city, country = "", dense = 0.2)
}
\arguments{
\item{api}{api of user}

\item{city}{name of a city lowercase}

\item{dense}{Transparency of points on plot}
}
\value{
A 2 by 2 plot latitude and longitude plot of points where the transparancy of the points corresponds to how cloudy that area is
}
\description{
Returns a plot of how dense clouds are in a given area
}

