% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/country_concordance.R
\docType{data}
\name{country_concordance}
\alias{country_concordance}
\title{Country concordance data.}
\format{
A data frame with codes as columns.
}
\description{
Country concordance of NOAA's not-harmonized two digit country code to country names and ISO 3 country codes. In NOAA's nomenclature `ES` is El Salvador, while the ISO 2 code for El Salvador is `SV`, while Spain has `ES`.
}
\details{
## Columns

- `NOAA_ccode`: NOAA's two digit country code
- `country_name`: country name (English)
- `country_code`: ISO-3 character
}
\note{
Some `NOAA_ccode`'s could not be matched by the Google API. They have `NA` values for `country_name` and `country_code`
}
\examples{
#> load country concordance

country_concordance

#>     NOAA_ccode                      country_name country_code
#> 1:         AA                             Aruba          ABW
#> 2:         AC               Antigua and Barbuda          ATG
#> 3:         AE  Abu Dhabi - United Arab Emirates          ARE
#> 4:         AF                              <NA>         <NA>
#> 5:         AG                           Algeria          DZA
#> ---                                                          
#> 246:         WS                 Western Sahara          ESH
#> 247:         WZ                       Eswatini          SWZ
#> 248:         YM                           Oman          OMN
#> 249:         ZA                         Zambia          ZMB
#> 250:         ZI                       Zimbabwe          ZWE

}
\keyword{datasets}
