% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_cover_crop_cardinal_temperatures.R
\docType{data}
\name{cover_crop_cardinal_temperatures}
\alias{cover_crop_cardinal_temperatures}
\title{cover_crop_cardinal_temperatures}
\format{
A data frame with 8 rows and 5 variables:
\describe{
\item{ cover_crop }{  character }
\item{ T_base }{  numeric }
\item{ T_opt }{  numeric }
\item{ T_max }{  numeric }
\item{ DOI }{  character }
}
}
\source{
10.1002/agg2.20393
}
\usage{
cover_crop_cardinal_temperatures
}
\description{
Base, optimum and maximum temperatures.
}
\keyword{datasets}
