% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwd.R
\name{cwd}
\alias{cwd}
\title{Get current weather description (cwd)}
\usage{
cwd(key, location)
}
\arguments{
\item{key}{A string, your API key}

\item{location}{A string, the location where you want to check the weather}
}
\value{
A string that describe current weather in the location indicated
}
\description{
The function helps to extract description of current weather from the OpenWeatherMap API
}
\examples{

key<-"&APPID=8196eccade1594d99afd30e0924dca7d"
cwd(key,"Vancouver,CA")

}
