% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distmap.R
\name{distmap}
\alias{distmap}
\title{distmap(api, city1, city2)}
\usage{
distmap(api, city1, city2)
}
\arguments{
\item{api}{api of user}

\item{city1}{name of a city lowercase}

\item{city2}{name of a city lowercase}
}
\value{
Distance between two cities in km
}
\description{
Returns distance between two cities.
}

