% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_weather_data.R
\name{download_weather_data}
\alias{download_weather_data}
\title{Download weather data from NOAA}
\usage{
download_weather_data(usaf, wban, year, dir)
}
\arguments{
\item{usaf}{USAF identifier of the weather station. Togehter with the `wban` identifier, this uniquely identifies the weather station.}

\item{wban}{WBAN identifier of the weather station. Togehter with the `usaf` identifier, this uniquely identifies the weather station.}

\item{year}{Year of the weather data.}

\item{dir}{Directory to save the weather data files.}
}
\description{
Downloads weather data from [NOAA](https://www.noaa.gov ). The data is downloaded in a compressed `gz` format.
}
