% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_evapotranspiration_fao56.R
\name{e_evapotranspiration_fao56}
\alias{e_evapotranspiration_fao56}
\title{Evapotranspiration de reference}
\source{
Irrigation and drainage paper, Allen & all (1998)
}
\usage{
e_evapotranspiration_fao56(
  altitude,
  R_inc,
  T_air_moy,
  T_air_min,
  T_air_max,
  vent = 2,
  pression_vapeur = z_pression_vapeur_saturante(T_air_min),
  a = NA,
  b = NA
)
}
\arguments{
\item{altitude}{Altitude de la localisation d'etude (m)}

\item{R_inc}{rayonnement incident du jour (MJ/m^2/j)}

\item{T_air_moy}{temperature moyenne du jour (celcius)}

\item{T_air_min}{temperature minimale du jour (celcius)}

\item{T_air_max}{temperature maximale du jour (celcius)}

\item{vent}{vitesse du vent a 2 m (m/s). Valeur par defaut egale a 2 m/s
d'apres p63 Allan & all. 1998.}

\item{pression_vapeur}{calcul interne a partir de la temperature minimale :
T_point_rosee = T_air_min d'apres Allan & all. 1998.}

\item{a}{Parametre correctif facultatif de Evapotranspiration corrige = a . Evapotranspiration + b.}

\item{b}{Parametre correctif facultatif de Evapotranspiration corrige = a . Evapotranspiration + b.}
}
\value{
Evapotranspiration de reference (mm)
}
\description{
Evapotranspiration de reference
}
\examples{
# Without correction
e_evapotranspiration_fao56(altitude = 30,
                           R_inc = 5,
                           T_air_moy = 20,
                           T_air_min = 10,
                           T_air_max = 30)

# With correction
e_evapotranspiration_fao56(altitude = 30,
                           R_inc = 5,
                           T_air_moy = 20,
                           T_air_min = 10,
                           T_air_max = 30,
                           a = 1.2,
                           b = 2)
}
