% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_weather_four_days.R
\name{get_weather_four_days}
\alias{get_weather_four_days}
\title{Retrieve 4-day weather forecast}
\usage{
get_weather_four_days(appid, city)
}
\arguments{
\item{appid}{API Key for openweathermap.org. Get one here: http://openweathermap.org/appid}

\item{city}{City name and country code divided by comma, use ISO 3166 country codes}
}
\value{
Data frame containing temp_min, temp_max, humidity and weather for the next 4 days in city local time. Temperature in Celsius
}
\description{
Retrieve 4-day weather forecast
}
