% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/getStationByID.R
\name{getStationByID}
\alias{getStationByID}
\title{Get ISD station data by USAFID for a range of years}
\usage{
getStationByID(stationID, station.list, begin, end)
}
\arguments{
\item{stationID}{USAFID station number}

\item{station.list}{Full list of ISD stations included in the package}

\item{begin}{Starting *year* to download data (4 digit)}

\item{end}{Ending *year* to download data (4 digit)}
}
\value{
Returns a list of two items.
1) Status of downloading each year.
2) Weather data in a list
}
\description{
Get ISD station data by USAFID for a range of years
}

