% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/getStationsByCity.R
\name{getStationsByCity}
\alias{getStationsByCity}
\title{Get k-nearest station data}
\usage{
getStationsByCity(city.list, station.list, k = 5, begin, end)
}
\arguments{
\item{city.list}{City of list of Cities. The format should be as follows: "City, State", or "City, Country"}

\item{station.list}{Full list of ISD stations included in the package}

\item{k}{The number of stations to return}

\item{begin}{Start year (4 digits)}

\item{end}{End year (4 digits)}
}
\value{
Returns a list of two items.

    1) Status of downloading each year's data for each station
    2) A list of dataframes.  Each dataframe is all years data of a particular station.
}
\description{
This function utilizes the kNStation() function to find the k-nearest stations,
and downloads all the data for those stations in between a given date range (in years).
}
\examples{
\dontrun{

data(stations) #called in as 'station.list'
cities <- c("Nairobi, Kenya", "Tema, Ghana", "Accra, Ghana", "Abidjan, Ivory Coast")
get.stations <- getStationsByCity(cities, station.list, begin = 2012, end = 2013)

get.stations$dl_status

class(get.stations$station_data)
length(get.stations$station_data)
}
}

