% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{match_api}
\alias{match_api}
\title{Match API icon to Weather Icon}
\usage{
match_api(
  name,
  api = c("darksky", "forecast-io", "yahoo", "owm", "wu", "wmo4680")
)
}
\arguments{
\item{name}{Name of the icon to find.}

\item{api}{Name of the API.}
}
\value{
a character vector.
}
\description{
Match API icon to Weather Icon
}
\examples{
# Default with darksky API
match_api("clear-day")
weather_icon(match_api("clear-day"))
}
