% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_nasapower_radiation_spindletop.R
\docType{data}
\name{nasapower_radiation_spindletop}
\alias{nasapower_radiation_spindletop}
\title{nasapower_radiation_spindletop}
\format{
A data frame with 2557 rows and 2 variables:
\describe{
\item{ date }{  Date }
\item{ radiation }{  numeric }
}
}
\source{
Source
}
\usage{
nasapower_radiation_spindletop
}
\description{
Observations of all sky solar radiation (PAR) from NASA POWER obtained at the coordiantes of the Landscape Nitrogen Position fields..
}
\keyword{datasets}
