% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{new_hi_lo_temp}
\alias{new_hi_lo_temp}
\title{Constructor for high/low temperature class}
\usage{
new_hi_lo_temp(highs, lows, dates, location, unit)
}
\arguments{
\item{highs}{Numeric vector of high temperatures}

\item{lows}{Numeric vector of low temperatures}

\item{dates}{Numeric vector of high temperatures}

\item{location}{Character vector containing city, state, and zip code}

\item{unit}{Unit that the temperature is measured in (Fahrenheit, Celsius, or Kelvin)}
}
\value{
An object of class hi_lo_temp
}
\description{
Creates a class that holds vectors for the high and low temperatures for a series of dates at a particular location
}
