% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipitation.R
\name{new_precip_data}
\alias{new_precip_data}
\title{Constructor for the Precipitation class}
\usage{
new_precip_data(precipitation, dates, location)
}
\arguments{
\item{precipitation}{Numeric vector of precipitation amounts}

\item{dates}{Numeric vector of dates}

\item{location}{Character vector containing city, state, and zip code}
}
\value{
<<<<<<< Updated upstream
An object of class precipitation
=======
An object of class precip_class
>>>>>>> Stashed changes
}
\description{
Creates a class that contains vectors for the precipitation for a series of dates at a particular location.
}
