% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_data_class.R
\name{new_weather_data}
\alias{new_weather_data}
\title{Constructor for \code{weather_data} class}
\usage{
new_weather_data(df)
}
\arguments{
\item{df}{The data frame that holds the data}
}
\value{
An object of class weather_data
}
\description{
Creates a class that holds vectors for the high and low temperatures, precipitation,
and wind speeds for a series of dates at a given location
}
\examples{
weather_data(data.frame("Date" = Sys.Date(), "High" = 70, "Low" = 55, "Wind Speed" = 5, "Precipitation" = 0, City = "Northampton", State = "Massachusetts"))
}
