% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_noho_forecast.R
\name{noho_weather_trim}
\alias{noho_weather_trim}
\title{Trim Northampton weather data}
\usage{
noho_weather_trim()
}
\value{
Returns a data frame with the following columns:
\itemize{
\item City (character): the city of weather origin
\item State (character): the state of weather origin
\item Date (Date): a Date formatted version of Day
\item High (numeric): the peak temperature predicted to be reached
\item Low (numeric): the coolest the day is predicted to be
\item Wind Speed (numeric): how fast the wind is blowing
}
}
\description{
Changes the data frame from \code{\link[=weather_noho]{weather_noho()}} to have different columns
}
\examples{
trimmed <- noho_weather_trim()
print(trimmed)
}
