% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_parameters.R
\docType{data}
\name{parameters}
\alias{parameters}
\title{parameters}
\format{
A data frame with  rows and  variables:
\describe{
\item{ inch_to_mm }{  numeric }
\item{ mm_of_snow_to_mm_of_rain_usa }{  numeric }
\item{ y_scale_ombrothermic_diagram }{  numeric }
}
}
\source{
Source
}
\usage{
parameters
}
\description{
List of parameters used in the package..
}
\keyword{datasets}
