% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paris}
\alias{paris}
\title{Fake data about temperature in Paris}
\format{A \code{data.frame} with 24 rows and 3 variables:
\describe{
  \item{hour}{Timestamp}
  \item{temperature}{Temperature in celsius degree}
  \item{weather}{Weather icons names}
}}
\description{
Fake data about temperature in Paris
}
\keyword{data}
