% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{past_days}
\alias{past_days}
\title{Retrieve high and low temperature data}
\source{
https://www.localconditions.com/
}
\usage{
past_days(city = "northampton", state = "massachusetts", zip = "01060")
}
\arguments{
\item{city}{A character vector of the city name}

\item{state}{A character vector of the state name}

\item{zip}{A character vector of the zip code of the city}
}
\value{
A data frame of weather from the past 30 days
}
\description{
Scrapes www.localconditions.com
}
\examples{
past_days("northampton", "massachusetts", "01060")
}
