% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipitation.R
\name{past_precipitation}
\alias{past_precipitation}
\title{Retrieve past precipitation data}
\usage{
past_precipitation(city = "new-york", state = "new-york", zip = "10001")
}
\arguments{
\item{city}{A character vector of the city name}

\item{state}{A character vector of the state name}

\item{zip}{A character vector of the zip code of the city}
}
\value{
An object of class precipitation
}
\description{
Scrapes www.localconditions.com for the past precipitation data
}
