% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{plot.hi_lo_temp}
\alias{plot.hi_lo_temp}
\title{Graph high and low temperatures}
\usage{
\method{plot}{hi_lo_temp}(obj)
}
\arguments{
\item{obj}{A \code{\link{high_low_temp}} object}
}
\description{
Graphs high and low temperatures from the past several days
}
