% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipitation.R
\name{plot.precipitation}
\alias{plot.precipitation}
\title{Plot past precipitation}
\usage{
\method{plot}{precipitation}(pre_obj)
}
\arguments{
\item{pre_obj}{An object of class precipitation}
}
\description{
Graphs past precipitation for the given dates
}
