% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_noho_forecast.R
\name{plot_noho_weather}
\alias{plot_noho_weather}
\title{Plot Northampton weather}
\usage{
plot_noho_weather(x, y)
}
\arguments{
\item{x, y}{character vectors containing column names from \code{\link[=weather_noho]{weather_noho()}} for the axis}
}
\description{
Plot a given column from the data frame returned from \code{\link[=weather_noho]{weather_noho()}}
}
