% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/plotDailyMax.R
\name{plotDailyMax}
\alias{plotDailyMax}
\title{Plot Daily Max Temp for Multiple Cities}
\usage{
plotDailyMax(hourlyDF)
}
\arguments{
\item{hourlyDF}{Must be a dataframe with columns for TEMP, city, YR, M, and D}
}
\value{
Plots the daily max temperateu of one or multiple cities
}
\description{
This function requires that the dataframe fed into the funciton have columns named 'TEMP' for temperature,
'city' for the city or identifier, 'YR' for the year, 'M' for month, and 'D' for the day.
}
\examples{
\dontrun{
data(stations)
cities <- c("Nairobi, Kenya", "Tema, Ghana", "Accra, Ghana", "Abidjan, Ivory Coast")
hourly.data <- getInterpolatedDataByCity(cities, station.list, 5, 2010, 2013, 100, 3, .05)
plotDailyMax(hourly.data)
}
}

