% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/plotStations.R
\name{plotStations}
\alias{plotStations}
\title{Plot k-nearest stations}
\usage{
plotStations(city.list, station.list, k = 5)
}
\arguments{
\item{city.list}{City of list of Cities. The format should be as follows: "City, State", or "City, Country"}

\item{station.list}{Full list of ISD stations included in the package}

\item{k}{The number of stations to return}
}
\value{
This will produce single, or multple plots for each city
}
\description{
This function will plot each city's reference point, along with the
k-nearest stations to that point.
}
\examples{
\dontrun{

data(stations) #called in as 'station.list'
cities <- c("Nairobi, Kenya", "Tema, Ghana", "Accra, Ghana", "Abidjan, Ivory Coast")
plotStations(cities, station.list, 5)
}
}

