% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipitation.R
\name{precip_class}
\alias{precip_class}
\title{Precipitation class}
\usage{
precip_class(precipitation, dates, location)
}
\arguments{
\item{precipitation}{Numeric vector of precipitation inches}

\item{dates}{Numeric vector of dates}

\item{location}{Character vector containing city, state, and zip code}
}
\value{
An object of class precip_class
}
\description{
A class that contains vectors for the precipitation for a series of dates at a particular location
}
