% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData.R
\name{prepareData}
\alias{prepareData}
\title{Title}
\usage{
prepareData(data)
}
\arguments{
\item{data}{}
}
\value{
A data frame.
}
\description{
Title
}
\examples{
prepareData(read.csv(file="wind-data.csv"))
}
