% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_radiation_inc_fao56.R
\name{r_radiation_inc_fao56}
\alias{r_radiation_inc_fao56}
\title{Calcul du rayonnement incident net}
\source{
Irrigation and drainage paper, Allen & all (1998)
}
\usage{
r_radiation_inc_fao56(
  longitude,
  latitude,
  altitude,
  date,
  T_air_min,
  T_air_max,
  albedo = 0.23,
  a = NA,
  b = NA
)
}
\arguments{
\item{longitude}{Longitude de la localisation d'etude (°)}

\item{latitude}{Latitude de la localisation d'etude (°)}

\item{altitude}{Altitude de la localisation d'etude (m)}

\item{date}{date au format "Y-m-d"}

\item{T_air_min}{temperature minimale du jour (celcius)}

\item{T_air_max}{temperature maximale du jour (celcius)}

\item{albedo}{albedo du gazon hypothetique de reference egale a 0.23 Eq. 38
Allan & all. 1998.}

\item{a}{Parametre correctif facultatif de l'equation Rayonnement net corrige = a . Rayonemment net + b.}

\item{b}{Parametre correctif facultatif de l'equation Rayonnement net corrige = a . Rayonemment net + b.}
}
\value{
Rayonnement net (MJ/m^2)
}
\description{
Calcul du rayonnement incident net
}
\examples{
# Without correction
r_radiation_inc_fao56(longitude = 10,
                      latitude = 35,
                      altitude = 30,
                      date = "2022-11-04",
                      T_air_min = 5,
                      T_air_max = 15)

# With correction
r_radiation_inc_fao56(longitude = 10,
                      latitude = 35,
                      altitude = 30,
                      date = "2022-11-04",
                      T_air_min = 5,
                      T_air_max = 15,
                      a = 1.2,
                      b = 2)
}
