% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_weather_data.R
\name{read_weather_data}
\alias{read_weather_data}
\title{Read weather data}
\usage{
read_weather_data(file)
}
\arguments{
\item{file}{Path to the weather data file.}
}
\value{
data frame of weather data.
}
\description{
`read_weather_data()` Reads weather data from a file.
}
