% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_daily_thermal_unit.R
\name{t_daily_thermal_unit}
\alias{t_daily_thermal_unit}
\title{w_daily_thermal_unit}
\usage{
t_daily_thermal_unit(
  T_air_min,
  T_air_max,
  T_base,
  T_opt = NA,
  T_max = NA,
  method = 1
)
}
\arguments{
\item{T_air_min}{Minimum air temperature (°C)}

\item{T_air_max}{Maximum air temperature (°C)}

\item{T_base}{Base air temperature of a crop (°C)}

\item{T_opt}{Optimum air temperature of a crop (°C)}

\item{T_max}{Maximum air temperature of a crop (°C)}

\item{method}{Selection of method 1 or 2. NOT the third one.}
}
\value{
Daily thermal unit (°Cj)
}
\description{
Calculation of the daily thermal unit.
Two methods are available.
\itemize{
\item Method 1: use of T_min, T_max and T_base. Method widely used for
the calculation of the daily thermal unit of small grain cereals.
\item Method 1: use of T_min, T_max and T_base. Method widely used for
the calculation of the daily thermal unit of corn.
\item Method 2: use of T_min, T_max, T_base, T_opt and T_max
}
}
\examples{
# Method 1
t_daily_thermal_unit(T_air_min = 5,
                     T_air_max = 8,
                     T_base = 10,
                     method = 1)

# Method 2
t_daily_thermal_unit(T_air_min = 5,
                     T_air_max = 8,
                     T_base = 10,
                     method = 2)

# Method 3
t_daily_thermal_unit(T_air_min = 5,
                     T_air_max = 8,
                     T_base = 10,
                     T_opt = 10,
                     T_max = 20)
}
