% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_fahrenheit_celcius.R
\name{t_fahrenheit_celcius}
\alias{t_fahrenheit_celcius}
\title{t_fahrenheit_celcius}
\usage{
t_fahrenheit_celcius(T_air_celcius = NA, T_air_fahrenheit = NA)
}
\arguments{
\item{T_air_celcius}{Air temperature (°C).}

\item{T_air_fahrenheit}{Air temperature (°F).}
}
\value{
Converted air temperature.
}
\description{
Conversion of degree Fahrenheit intp degree celcius and
vice-versa.
}
\examples{
# From fahrenheit to celcius
t_fahrenheit_celcius(T_air_fahrenheit = 72)

# From celcius to fahrenheit
t_fahrenheit_celcius(T_air_celcius = 22)
}
