% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{temps_today}
\alias{temps_today}
\title{Retrieve today's temperatures}
\usage{
temps_today(city = "northampton", state = "massachusetts", zip = "01060")
}
\value{
Numeric vectors containing the past day's temperatures
}
\description{
Returns approximately hourly temperatures over the course of the current day
}
