% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_data_class.R
\name{to_hi_lo_temp}
\alias{to_hi_lo_temp}
\title{Create \code{hi_lo_temp} from \code{weather_data}}
\usage{
to_hi_lo_temp(weather_data)
}
\arguments{
\item{weather_data}{An object of class \code{weather_data}}
}
\value{
An object of class \code{hi_lo_temp}
}
\description{
Creates an object of high and low temperatures from weather data
}
\examples{
to_hi_lo_temp(weather_data(data.frame("Date" = Sys.Date(), "High" = 70, "Low" = 55, "Wind Speed" = 5, "Precipitation" = 0, City = "Northampton", State = "Massachusetts")))
}
