% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_data_class.R
\name{to_precipitation}
\alias{to_precipitation}
\title{Create \code{precipitation} from \code{weather_data}}
\usage{
to_precipitation(weather_data)
}
\arguments{
\item{weather_data}{An object of class \code{weather_data}}
}
\value{
An object of class \code{precipitation}
}
\description{
Creates an object of precipitation amount from past weather data
}
\examples{
to_precipitation(weather_data(data.frame("Date" = Sys.Date(), "High" = 70, "Low" = 55, "Wind Speed" = 5, "Precipitation" = 0, City = "Northampton", State = "Massachusetts")))
}
