% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USweather.R
\docType{data}
\name{USweather}
\alias{USweather}
\title{Weather Data for US Cities (2016)}
\format{
\describe{
\item{Precipitation}{The average amount of rain, in inches.}
\item{Date}{A full string representation of the date this report was made.}
\item{Month}{The month of the year that this report was made.}
\item{Week}{The first day of the week that this report was made.}
\item{Year}{The year that this report was made.}
\item{City}{The city that the reporting station sends its data to.}
\item{Station Code}{The unique code representing this station.}
\item{Location}{The exact location of this recording station.}
\item{State}{The state that the reporting station sends its data to.}
\item{Avg. Temp}{The average recorded temperature on this week, in degrees Fahrenheit.}
\item{High}{The highest recorded temperature on this week, in degrees Fahrenheit.}
\item{Low}{The lowest recorded temperature on this week, in degrees Fahrenheit.}
\item{Wind Direction}{The average wind direction for that week, in degrees.}
\item{Wind Speed}{The average windspeed for that week, in Miles per Hour.}
}
}
\source{
\url{http://w2.weather.gov/climate/}
}
\usage{
USweather
}
\description{
Weather Data for US Cities (2016)
}
\keyword{datasets}
