% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hi_lo_temp_class.R
\name{validate_hi_lo_temp}
\alias{validate_hi_lo_temp}
\title{Validator for high/low temperature class}
\usage{
validate_hi_lo_temp(obj)
}
\arguments{
\item{obj}{A \code{\link[=hi_lo_temp]{hi_lo_temp()}} object to be validated}
}
\value{
An object of class hi_lo_temp
}
\description{
Validates a class that holds vectors for the high and low temperatures for a series
of dates at a particular location
}
