% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipitation.R
\name{validate_precip_data}
\alias{validate_precip_data}
\title{Validator for Precipitation class}
\usage{
validate_precip_data(pre_obj)
}
\arguments{
<<<<<<< Updated upstream
\item{obj}{Object of class precipitation}
}
\value{
An object of class precipitation
=======
\item{obj}{Object of class precip_class}
}
\value{
An object of class precip_class
>>>>>>> Stashed changes
}
\description{
Validates a class that holds vectors for the precipitation data for a series of dates at a particular location
}
