% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_data_class.R
\name{validate_weather_data}
\alias{validate_weather_data}
\title{Validator for \code{weather_data} class}
\usage{
validate_weather_data(obj)
}
\arguments{
\item{obj}{An object of class weather_data}
}
\value{
An object of class weather_data
}
\description{
Validates a class that holds high and low temperatures, precipitation, and wind
speed data for a series of dates at a particular location
}
\examples{
validate_weather_data(new_weather_data(weather_data(data.frame("Date" = Sys.Date(),
"High" = 70, "Low" = 55, "Wind Speed" = 5, "Precipitation" = 0, City = "Northampton",
State = "Massachusetts"))))
}
