% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{weather-dependencies}
\alias{weather-dependencies}
\alias{use_weather_icons}
\alias{weather_dependencies}
\title{Weather Icons Dependencies}
\usage{
use_weather_icons()

weather_dependencies()
}
\description{
Explicitly load weathr icons dependencies in shiny or markdown.
}
