\name{weather-package}
\alias{weather-package}
\alias{weather}
\docType{package}

\title{
What the package does (short line)
	Weather data manipulation
}


\description{
More about what it does (maybe more than one line)
	Weather data manipulation. Particularly in the context of ecological modeling, such as crop growth simulation modeling.
}

\details{
An overview of how to use the package, including the most important functions ~~
}


\keyword{ package }
\keyword{ spatial }

