% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_weather_blue_grass_airport.R
\docType{data}
\name{weather_blue_grass_airport}
\alias{weather_blue_grass_airport}
\title{weather_blue_grass_airport}
\format{
A data frame with 12419 rows and 9 variables:
\describe{
\item{ date }{  Date }
\item{ year }{  numeric }
\item{ month }{  numeric }
\item{ day }{  integer }
\item{ rain }{  numeric }
\item{ snow }{  numeric }
\item{ T_air_max }{  numeric }
\item{ T_air_min }{  numeric }
\item{ T_air_avg }{  numeric }
}
}
\source{
Source
}
\usage{
weather_blue_grass_airport
}
\description{
Observations of the blue grass airport weather station.
}
\keyword{datasets}
