% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_data_class.R
\name{weather_data}
\alias{weather_data}
\title{Weather data class}
\usage{
weather_data(df)
}
\arguments{
\item{df}{The data frame that holds the data}
}
\value{
An object of the class weather_data
}
\description{
A class that holds high and low temperature, wind speed, precipitation, and location for given dates
}
\examples{
weather_data(data.frame("Date" = Sys.Date(), "High" = 70, "Low" = 55, "Wind Speed" = 5, "Precipitation" = 0, City = "Northampton", State = "Massachusetts"))
}
