% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weather_icon_names}
\alias{weather_icon_names}
\title{Weather Icon Names (for ggplot2)}
\format{A \code{character} \code{vector}.}
\description{
Weather Icon Names (for ggplot2)
}
\keyword{data}
