% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noho_forecast.R
\name{weather_noho}
\alias{weather_noho}
\title{Get data from weather.com}
\source{
\url{https://weather.com/weather/tenday/l/8a8df2c2297bffb63bf3fc70a45fd88e10152b88363a57940cb501eb7d2f8094}
}
\usage{
weather_noho()
}
\value{
Returns a data frame with the following columns:
\itemize{
\item Day (character): day of the week and the day of the month
\item Date (Date): a Date formatted version of Day
\item Status (character): conditions outside (ex. rainy, cloudy, sunny)
\item High (numeric): the peak temperature predicted to be reached
\item Low (numeric): the coolest the day is predicted to be
\item Precipitation Type (character): what type of precipitation (ex. rain, snow)
\item Precipitation Chance (numeric): chances that it will rain, snow, etc
\item Wind Direction (character): the direction that the wind is blowing
\item Wind Speed (numeric): how fast the wind is blowing
}
}
\description{
Get data from weather.com about Northampton, MA weather for the next 14 days
}
\examples{
forecast <- weather_noho()
}
