% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_angle_conversion.R
\name{z_angle_conversion}
\alias{z_angle_conversion}
\title{Conversion des angles en degres en radians et vice versa}
\usage{
z_angle_conversion(angle, mode)
}
\arguments{
\item{angle}{valeur de l'angle}

\item{mode}{1 : conversion des degres en radians 2 : conversion des radians en degres}
}
\value{
valeur convertie de l'angle
}
\description{
Conversion des angles en degres en radians et vice versa
}
\examples{
z_angle_conversion(pi, 2)
z_angle_conversion(180, 1)
}
