% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_pression_atm.R
\name{z_pression_atm}
\alias{z_pression_atm}
\title{Calcul de la pression atmospherique en fonction de l'altitude de la station}
\usage{
z_pression_atm(altitude)
}
\arguments{
\item{altitude}{elevatation au dessus du niveau de la mer (m)}
}
\value{
pression atmospherique (kPa)
}
\description{
Calcul de la pression atmospherique en fonction de l'altitude de la station
}
\examples{
z_pression_atm(0)
z_pression_atm(5000)
}
