% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_pression_vapeur_saturante.R
\name{z_pression_vapeur_saturante}
\alias{z_pression_vapeur_saturante}
\title{Calcul de la pression de vapeur saturante pour une temperature donnee}
\usage{
z_pression_vapeur_saturante(temperature)
}
\arguments{
\item{temperature}{temperature de l'air (degre celcius)}
}
\value{
pression de vapeur saturante (kPa)
}
\description{
Calcul de la pression de vapeur saturante pour une temperature donnee
}
\examples{
z_pression_vapeur_saturante(20)
}
