% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_weather_stations.R
\name{available_weather_stations}
\alias{available_weather_stations}
\title{Download available weather stations}
\usage{
available_weather_stations(file_dir = "", save_file = TRUE)
}
\arguments{
\item{file_dir}{Directory to save the CSV file. If empty, a temporary directory is created.}

\item{save_file}{Logical indicating whether to save the CSV file.}
}
\value{
data frame of available weather stations.
}
\description{
`available_weather_stations()` Downloads all available weather stations from NOAA.
}
