% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_weather_station.R
\name{closest_weather_station}
\alias{closest_weather_station}
\title{Find the closest weather station(s) to a given location}
\usage{
closest_weather_station(latitude, longitude, stations, n = 1)
}
\arguments{
\item{latitude}{Latitude of the location.}

\item{longitude}{Longitude of the location.}

\item{stations}{Data frame of weather stations. Use `available_weather_stations()` to get all available weather stations.}

\item{n}{Number of closest weather stations to return.}
}
\value{
data frame of closest weather station(s).
}
\description{
`closest_weather_station()` Finds the closest weather station(s) to a given location.
}
