% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_weather_data.R
\name{download_weather_data}
\alias{download_weather_data}
\title{Download weather data from NOAA}
\usage{
download_weather_data(usaf, wban, begin_year, end_year, dict)
}
\arguments{
\item{usaf}{USAF identifier of the weather station.}

\item{wban}{WBAN identifier of the weather station.}

\item{begin_year}{Beginning year of the weather data.}

\item{end_year}{Ending year of the weather data.}

\item{dict}{Directory to save the weather data files.}
}
\description{
`download_weather_data()` Downloads weather data from NOAA.
}
