% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hi_lo_temp_class.R
\name{hi_lo_temp}
\alias{hi_lo_temp}
\title{High/low temperature class}
\usage{
hi_lo_temp(highs, lows, dates, location)
}
\arguments{
\item{highs}{Numeric vector of high temperatures}

\item{lows}{Numeric vector of low temperatures}

\item{dates}{Numeric vector of high temperatures}

\item{location}{Character vector containing city, state, and zip code}
}
\value{
An object of class hi_lo_temp
}
\description{
A class that holds vectors for the high and low temperatures for a series of dates
at a particular location
}
\examples{
hi_lo_temp(80, 60, Sys.Date(), c("Northampton", "MA"))
}
