% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/kNStations.R
\name{kNStations}
\alias{kNStations}
\title{Get the k-nearest weather stations to a city}
\usage{
kNStations(city.list, station.list, k = 5)
}
\arguments{
\item{city.list}{List of cities. Format should be: "City, State", or "City, Country"}

\item{station.list}{Full list of ISD stations included in the package}

\item{k}{The number of stations to return}
}
\value{
Returns the k-nearest weather stations in the full ISD station.list
}
\description{
Using either an individual city or a list of cities, this will find a reference point for each location
and return the k-nearest weather stations to that reference point.
}
\examples{
\dontrun{
data(stations)
cities <- c("Nairobi, Kenya", "Tema, Ghana", "Accra, Ghana", "Abidjan, Ivory Coast")
k.n.stations <- kNStations(cities, station.list, 5)
}
}

