% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{past_days}
\alias{past_days}
\title{Retrieve high and low temperature data}
\usage{
past_days(city = "northampton", state = "massachusetts", zip = "01060")
}
\arguments{
\item{city}{A character vector of the city name}

\item{state}{A character vector of the state name}

\item{zip}{A character vector of the zip code of the city}
}
\value{
An object of class \code{\link{hi_lo_temp}}
}
\description{
Scrapes www.localconditions.com
}
