% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipitation.R
\name{past_precipitation}
\alias{past_precipitation}
\title{Retrieve past precipitation data}
<<<<<<< Updated upstream
=======
\source{
\url{https://www.localconditions.com/}
}
>>>>>>> Stashed changes
\usage{
past_precipitation(city = "new-york", state = "new-york", zip = "10001")
}
\arguments{
\item{city}{A character vector of the city name}

\item{state}{A character vector of the state name}

\item{zip}{A character vector of the zip code of the city}
}
\value{
<<<<<<< Updated upstream
An object of class precipitation
=======
An object of class precip_class
>>>>>>> Stashed changes
}
\description{
Scrapes www.localconditions.com for the past precipitation data
}
<<<<<<< Updated upstream
=======
\examples{
past_precipitation(city = "northampton", state = "massachusetts", zip = "01060")

}
>>>>>>> Stashed changes
