% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hi_lo_temp_class.R
\name{plot.hi_lo_temp}
\alias{plot.hi_lo_temp}
\title{Graph high and low temperatures}
\usage{
\method{plot}{hi_lo_temp}(x, ...)
}
\arguments{
\item{x}{A \code{\link[=hi_lo_temp]{hi_lo_temp()}} object}

\item{...}{Additional parameters}
}
\description{
Graphs high and low temperatures from the past several days
}
\examples{
plot(hi_lo_temp(c(80,70), c(60,65), seq(Sys.Date()-1, Sys.Date(), , by="days"), c("Northampton", "MA")))
}
