% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precipitation_class.R
\name{plot.precipitation}
\alias{plot.precipitation}
\title{Plot past precipitation}
\usage{
\method{plot}{precipitation}(x, ...)
}
\arguments{
\item{x}{An object of class precip_class generated from function past_precipitation}

\item{...}{Additional parameters}
}
\description{
Graphs past precipitation for the given dates
}
\examples{
pastprecip <- precipitation(c(0.01, 0, 0.04, 0),
seq(Sys.Date()-3, Sys.Date(), by="days"), c("Northampton","Massachusetts"))
plot(pastprecip)
}
