% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_noho_forecast.R
\name{plot.weather}
\alias{plot.weather}
\title{Plot Northampton weather}
\usage{
\method{plot}{weather}(data, x, y)
}
\arguments{
\item{data}{data frame of the weather class from \code{\link[=weather_noho]{weather_noho()}}}

\item{x, y}{character vectors containing column names from \code{\link[=weather_noho]{weather_noho()}} for the axis}
}
\description{
Create a scatterplot using two variables from a data frame from \code{\link[=weather_noho]{weather_noho()}}
}
\examples{
weather <- weather_noho()
plot(weather, "High", "Low")
}
