% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_weather_data.R
\name{read_weather_data}
\alias{read_weather_data}
\title{Read weather data}
\usage{
read_weather_data(file)
}
\arguments{
\item{file}{Path to the weather data file.}
}
\value{
data frame of weather data.

Returns a data.frame with 10 columns.
\describe{
\item{station_id}{Station ID of the weather station.}
\item{time}{Time of the weather observation in GMT timezone.}
\item{wd}{Wind direction in degrees.}
\item{ws}{Wind speed in meters per second.}
\item{ceil_hgt}{Ceiling height in meters.}
\item{visibility}{Visibility in meters.}
\item{temp}{Temperature in degrees Celsius.}
\item{dew_point}{Dew point in degrees Celsius.}
\item{atmos_pres}{Atmospheric pressure in hPa.}
\item{rh}{Relative humidity in percent.}
}
}
\description{
Reads weather data, obtained with `download_weather_data()`, from a file. Relative humidity is calculated from temperature and dew point.
}
\examples{
\dontrun{
# Read weather data from a file
data <- read_weather_data("data/037070-99999-2017.gz")
}

}
