% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_daily_thermal_unit.R
\name{t_daily_thermal_unit}
\alias{t_daily_thermal_unit}
\title{t_daily_thermal_unit}
\usage{
t_daily_thermal_unit(
  T_air_avg,
  T_base,
  T_opt = NA,
  T_opt_1 = NA,
  T_opt_2 = NA,
  T_max = NA,
  method = "A"
)
}
\arguments{
\item{T_air_avg}{Average air temperature (°C)}

\item{T_base}{Base air temperature of a crop (°C)}

\item{T_opt}{Optimum air temperature of a crop (°C)}

\item{T_opt_1}{First optimum temperature of a crop (°C)}

\item{T_opt_2}{Second optimum temperature of a crop (°C)}

\item{T_max}{Maximum air temperature of a crop (°C)}

\item{method}{Selection of the calculation method:
\itemize{
\item Method A: Linear - small grains
\item Method B: Linear - corn
\item Method C: Triangle
\item Method D: Trapezoidal
}}
}
\value{
Daily thermal unit (°Cj)
}
\description{
Calculation of the daily thermal unit.
Two methods are available.
\itemize{
\item Method A - linear - small grains: use of T_base.
Method widely used for the calculation of the daily thermal unit of small
grain cereals.
\item Method B - linear: use of T_air_min, T_air_max and T_base. Method
widely used for the calculation of the daily thermal unit of corn.
\item Method C - triangle: use of T_air_min, T_air_max, T_base, T_opt and
T_max.
\item Method D - trapezoid: use of T_air_min, T_air_max, T_base, T_opt_1,
T_opt_2 and T_max.
}
}
\examples{
# Method A - linear - small grains
t_daily_thermal_unit(T_air_avg = 15,
                     T_base = 5,
                     method = "A")

# Method B - linear - corn
t_daily_thermal_unit(T_air_avg = 15,
                     T_base = 5,
                     method = "B")

# Method C - triangle
t_daily_thermal_unit(T_air_avg = 15,
                     T_base = 5,
                     T_opt = 10,
                     T_max = 20,
                     method = "C")

# Method D - trapezoidal
t_daily_thermal_unit(T_air_avg = 15,
                     T_base = 5,
                     T_opt_1 = 12,
                     T_opt_2 = 18,
                     T_max = 20,
                     method = "D")
}
