% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/weatheR-package.R
\docType{package}
\name{weatheR-package}
\alias{weatheR}
\alias{weatheR-package}
\title{Download, Transform, and Plot NOAA ISD weather data}
\description{
This is an R package with all the necessary functions to download,
transform, and plot NOAA weather station data.
}
\author{
Michael Piccirilli \email{mrp2181@columbia.edu}
}
\seealso{
\code{\link{allStations}}, \code{\link{getFilteredStationsByCity}},
\code{\link{getInterpolatedDataByCity}}, \code{\link{getStationByID}},
\code{\link{getStationsByCity}}, \code{\link{kNStations}},
\code{\link{plotDailyMax}}, \code{\link{plotStations}}
}

